finished UI for entity attribute
