+++ modules/core/db/update/hsql/18/180515-2-updateImportAttributeMapper01.sql
@@ -1,2 +1,1 @@
-alter table DDCDI_IMPORT_ATTRIBUTE_MAPPER add column DYNAMIC_ENTITY_ATTRIBUTE varchar(255) ;
 alter table DDCDI_IMPORT_ATTRIBUTE_MAPPER add column ASSOCIATION_LOOKUP_ATTRIBUTE varchar(255) ;
