+++ modules/global/src/de/diedavids/cuba/dataimport/entity/messages.properties
@@ -31,7 +31,6 @@ ImportAttributeMapper.customAttributeBindScript = Custom attribute bind script
 ImportAttributeMapper.dynamicAttribute = Dynamic attribute
 ImportAttributeMapper.attributeType = Attribute type
 ImportAttributeMapper.associationLookupAttribute = Association lookup attribute
-ImportAttributeMapper.dynamicEntityAttribute = Dynamic entity attribute
 ImportConfiguration = Import Configuration
 ImportConfiguration.name = Configuration Name
 ImportConfiguration.entityClass = Entity class
