+++ modules/web/src/de/diedavids/cuba/dataimport/web/importattributemapper/messages.properties
@@ -1,6 +1,7 @@
 editorCaption = ImportAttributeMapper editor
 dataTab= Import Attribute Mapper settings
 customBindingScriptTab=Custom binding Script
+entityAttributeLabel = Entity Attribute
 customAttributeBindScriptHelp=This script can be used for defining custom binding logic. \n\
   The return value of this script will be set to the corresponding Entity attribute. \n\n\
   If this script is set, it will disable the auto-detection of the import process. Therefore it \
@@ -12,4 +13,10 @@ customAttributeBindScriptHelp=This script can be used for defining custom bindin
    - dataRow: the complete data row as it is taken from the import file \n\
    - entityAttribute: the current entity attribute \n\
    - importConfiguration: the current import configuration \n\
-   - importAttributeMapper: the current import attribute mapper entity \n\
+   - importAttributeMapper: the current import attribute mapper entity
+associationLookupAttributeHelpText=This attribute defines which column within the association attribute should \n\
+  be used for the Lookup for the existing entry in the database. \n\n\
+  It is also possible in case of multi-level reference to manually enter a value. The start point of this reference \n\
+  is the Entity selected in the Entity Attribute selection.
+entityAttributeHelpText=This selection defines which entity attribute will be used for importing the corresponding column. \
+  The values in this selection depend on the attribute type.
