+++ modules/web/src/de/diedavids/cuba/dataimport/web/importconfiguration/import-configuration-edit.xml
@@ -81,6 +81,7 @@
                     <columns>
                         <column id="fileColumnNumber"/>
                         <column id="fileColumnAlias"/>
+                        <column id="attributeType"/>
                         <column id="entityAttribute"/>
                     </columns>
                     <rows datasource="importAttributeMappersDs"/>
