+++ modules/web/src/de/diedavids/cuba/dataimport/web/importwizard/import-wizard.xml
@@ -100,6 +100,7 @@
                     <columns>
                         <column id="fileColumnNumber"/>
                         <column id="fileColumnAlias"/>
+                        <column id="attributeType"/>
                         <column id="entityAttribute"/>
                     </columns>
                     <buttonsPanel id="buttonsPanel"
